+++ Soundfingerprinting/IFingerprintDescriptor.cs
@@ -3,20 +3,20 @@ namespace Soundfingerprinting.Fingerprinting
 	public interface IFingerprintDescriptor
 	{
 		/// <summary>
-		///   Encode the integer representation of the fingerprint into Boolean array
+		/// Encode the integer representation of the fingerprint into Boolean array
 		/// </summary>
 		/// <param name = "concatenated">Concatenated fingerprint (frames concatenated)</param>
 		/// <param name = "indexes">Sorted indexes with the first one with the highest value in array</param>
 		/// <param name = "topWavelets">Number of top wavelets to encode</param>
 		/// <returns>Encoded fingerprint</returns>
-		bool[] EncodeFingerprint(float[] concatenated, int[] indexes, int topWavelets);
+		bool[] EncodeFingerprint(double[] concatenated, int[] indexes, int topWavelets);
 
 		/// <summary>
-		///   Decode the signature of the fingerprint
+		/// Decode the signature of the fingerprint
 		/// </summary>
 		/// <param name = "signature">Signature to be decoded</param>
 		/// <returns>Array of doubles with positive [10], negatives [01], and zeros [00]</returns>
-		float[] DecodeFingerprint(bool[] signature);
+		double[] DecodeFingerprint(bool[] signature);
 
 		/// <summary>
 		/// Sets all other wavelet values to 0 except whose which make part of Top Wavelet [top wavelet &gt; 0 ? 1 : -1]
@@ -35,6 +35,6 @@ namespace Soundfingerprinting.Fingerprinting
 		///   Positive Numbers = 10
 		///   Zeros            = 00
 		/// </remarks>
-		bool[] ExtractTopWavelets(float[][] frames, int topWavelets);
+		bool[] ExtractTopWavelets(double[][] frames, int topWavelets);
 	}
 }
